requirement.txt
alembic==1.0.6
baidu-aip==2.2.12.0
blueprint==3.4.2
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Migrate==2.3.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
mysqlclient==1.3.14
PyMySQL==0.9.3
python-dateutil==2.7.5
python-editor==1.0.3
qcloudsms-py==0.1.3
requests==2.21.0
six==1.12.0
SQLAlchemy==1.2.16
urllib3==1.24.1
Werkzeug==0.14.1
